refactor(accelerator): move the DuckDB accelerator into its own crate
Fourth engine out, and the one the runtime reached for in the most places.

`RuntimeBuilder` constructed a `DuckDBAccelerator` directly to resolve instance
identity for the coordinated memory budget, so `DataAccelerator` grows
`memory_budget_inputs` and the 108-line adapter moves into the engine. Keeping
that rule in one place is the point: instance identity resolves `duckdb_file`,
`duckdb_data_dir` and the default filename through the same factory the engine
opens databases with, and a second implementation in the planner could key
instances differently and mis-size every cap. The builder finds the engine
through the registration slice rather than naming its crate.

Three test targets and the Spicepod schema tool reached the module through
`runtime` and now name the crate. The schema tool's force-links lose their
`#[expect(unused_imports)]`: `use <crate> as _;` is not an unused import, so the
expectation goes unfulfilled, which `-Dwarnings` rejects — the module-path form
they replaced did need it.

`runtime-table` and `spice-table` are dev-dependencies here, not normal ones:
they sit in the `runtime` tier, and the cache-refresh tests that need them live
in this crate only because they need its `create_table_provider`.

The engine's own `#[cfg(feature = "duckdb")]` gates are gone — meaningless in a
crate that *is* the DuckDB engine — and its three insta snapshots move with it.
`runtime`'s `dialect` alias goes too, having lost its last user.
