+++ bin/spiced/src/lib.rs
@@ -106,6 +106,8 @@ use accelerator_postgres as _;
 use accelerator_sqlite as _;
 #[cfg(feature = "turso")]
 use accelerator_turso as _;
+#[cfg(feature = "duckdb")]
+use accelerator_duckdb as _;
 use connector_spiceai as _;
 use opentelemetry::{KeyValue, global};
 use opentelemetry_sdk::Resource;
