+++ crates/accelerators/accelerator-duckdb/src/settings.rs
