+++ crates/runtime/src/datafusion/mod.rs
@@ -34,7 +34,6 @@ use crate::component::view::View;
 use crate::dataaccelerator::ReloadProviderFactory;
 use crate::dataaccelerator::spice_sys::dataset_checkpointer;
 use crate::dataaccelerator::{self, BootstrapStatus};
-use runtime_acceleration::acceleration_source::resolved_refresh_mode;
 use crate::dataaccelerator::{AcceleratorEngineRegistry, get_acceleration_layout};
 use crate::dataconnector::deferred::DeferredConnector;
 use crate::dataconnector::localpod::LOCALPOD_DATACONNECTOR;
@@ -60,6 +59,7 @@ use crate::view::prepare_view;
 use crate::{status, view};
 use data_accelerator_api::swappable::SwappableTableProvider;
 use data_connector_api::federated::FederatedTableProvider;
+use runtime_acceleration::acceleration_source::resolved_refresh_mode;
 use runtime_acceleration::dataset_checkpoint::DatasetCheckpointer;
 use runtime_acceleration::sidecar::OpenOption;
 use runtime_acceleration::snapshot::SnapshotBehavior;
@@ -139,12 +139,10 @@ pub mod builder;
 #[cfg(not(windows))]
 pub mod cayenne_ddl;
 pub use runtime_datafusion::composed_catalog;
-// `dialect`, `error` and `refresh_sql` below are named throughout the runtime
-// through these aliases, but they belong to `runtime-datafusion`. Crate-visible
-// so a crate outside the runtime has to depend on `runtime-datafusion` directly
-// rather than route through here.
-#[cfg(any(feature = "duckdb", test))]
-pub(crate) use runtime_datafusion::dialect;
+// `error` and `refresh_sql` below are named throughout the runtime through these
+// aliases, but they belong to `runtime-datafusion`. Crate-visible so a crate outside
+// the runtime has to depend on `runtime-datafusion` directly rather than route
+// through here.
 pub(crate) use runtime_datafusion::error;
 pub use runtime_table::filter_converter;
 pub mod flight_session_extension;
