+++ tools/spicepodschema/src/collector.rs
@@ -73,14 +73,10 @@ use runtime::dataaccelerator::arrow as _;
 #[cfg(not(windows))]
 #[expect(unused_imports)]
 use runtime::dataaccelerator::cayenne as _;
-#[expect(unused_imports)]
-use runtime::dataaccelerator::duckdb as _;
-#[expect(unused_imports)]
-use runtime::dataaccelerator::postgres as _;
-#[expect(unused_imports)]
-use runtime::dataaccelerator::sqlite as _;
-#[expect(unused_imports)]
-use runtime::dataaccelerator::turso as _;
+use accelerator_duckdb as _;
+use accelerator_postgres as _;
+use accelerator_sqlite as _;
+use accelerator_turso as _;
 
 /// Schema information for a connector or accelerator.
 #[derive(Debug, Clone)]
